Add Map and AdjustCamera components; refactor Simulator for improved structure; fix Camera angle bugs periodically
