+++ src/components/viewControls.tsx
@@ -1,132 +1,48 @@
-import React from 'react';
+import { SimStoreState, useSimStore } from '@/store/useSimStore';
+import Image from 'next/image';
 
-interface ViewControlsProps {
-  onSetTopView: () => void;
-  onSetRotatedView: () => void;
-}
+export default function ViewControls() {
+  const { pos } = useSimStore((state) => state.view);
+  const setView = useSimStore((state) => state.setView);
 
-export default function ViewControls({
-  onSetTopView,
-  onSetRotatedView,
-}: ViewControlsProps) {
+  const onClick = (view?: SimStoreState['view']['pos']) => {
+    if (view === 'topView') {
+      setView({ pos: 'topView' });
+    } else {
+      setView({
+        pos: pos === 'rotatedSouth' ? 'rotatedNorth' : 'rotatedSouth',
+      });
+    }
+  };
+
   return (
-    <div
-      style={{
-        display: 'flex',
-        flexDirection: 'row',
-        justifyContent: 'center',
-        alignItems: 'center',
-        position: 'absolute',
-        zIndex: 10,
-        top: '5%',
-        left: '58%',
-        transform: 'translateX(-50%)',
-        background: 'rgba(255, 255, 255, 0.8)',
-        padding: '8px',
-        borderRadius: '8px',
-        boxShadow: '0px 4px 6px rgba(0, 0, 0, 0.1)',
-      }}
-    >
+    <div className='flex flex-row justify-center items-center absolute z-10 top-[5%] left-[58%] -translate-x-1/2 bg-white/80 p-2 rounded-lg shadow-md'>
       {/* Top View Button */}
       <button
-        onClick={onSetTopView}
-        style={{
-          display: 'flex',
-          alignItems: 'center',
-          justifyContent: 'center',
-          margin: '0 5px',
-          padding: '6px',
-          borderRadius: '8px',
-          border: 'none',
-          cursor: 'pointer',
-          background: 'rgb(73, 104, 118)',
-          color: '#fff',
-          flex: 1,
-          maxWidth: '170px',
-          minWidth: '170px',
-          textAlign: 'center',
-          transition: 'background-color 0.2s ease, transform 0.2s ease',
-          outline: 'none',
-          boxShadow: 'none',
-        }}
-        onMouseEnter={(e) => {
-          (e.target as HTMLElement).style.backgroundColor = 'rgb(45, 75, 90)';
-        }}
-        onMouseLeave={(e) => {
-          (e.target as HTMLElement).style.backgroundColor = 'rgb(73, 104, 118)';
-        }}
-        onMouseDown={(e) => {
-          (e.target as HTMLElement).style.backgroundColor = 'rgb(30, 60, 75)';
-          const img = (e.target as HTMLElement).querySelector('img');
-          if (img) img.style.transform = 'scale(1.1)';
-        }}
-        onMouseUp={(e) => {
-          (e.target as HTMLElement).style.backgroundColor = 'rgb(45, 75, 90)';
-          const img = (e.target as HTMLElement).querySelector('img');
-          if (img) img.style.transform = 'scale(1)';
-        }}
+        onClick={() => onClick('topView')}
+        className='flex items-center justify-center mx-[5px] p-[6px] rounded-lg border-none cursor-pointer bg-[rgb(73,104,118)] hover:bg-[rgb(45,75,90)] text-white flex-1 max-w-[170px] min-w-[170px] text-center transition-all outline-none shadow-none'
       >
-        <img
-          src="/assets/down_circle.svg"
-          alt="Top View Icon"
-          style={{
-            width: '18px',
-            height: '18px',
-            marginRight: '7px',
-            transition: 'transform 0.2s ease',
-          }}
+        <Image
+          src='/assets/down_circle.svg'
+          width={18}
+          height={18}
+          className='mr-[7px]'
+          alt='Top View Icon'
         />
         Top View
       </button>
 
       {/* Rotated View Button */}
       <button
-        onClick={onSetRotatedView}
-        style={{
-          display: 'flex',
-          alignItems: 'center',
-          justifyContent: 'center',
-          margin: '0 5px',
-          padding: '6px',
-          borderRadius: '8px',
-          border: 'none',
-          cursor: 'pointer',
-          background: 'rgb(73, 104, 118)',
-          color: '#fff',
-          flex: 1,
-          maxWidth: '170px',
-          minWidth: '170px',
-          textAlign: 'center',
-          transition: 'background-color 0.2s ease, transform 0.2s ease',
-          outline: 'none',
-          boxShadow: 'none',
-        }}
-        onMouseEnter={(e) => {
-          (e.target as HTMLElement).style.backgroundColor = 'rgb(45, 75, 90)';
-        }}
-        onMouseLeave={(e) => {
-          (e.target as HTMLElement).style.backgroundColor = 'rgb(73, 104, 118)';
-        }}
-        onMouseDown={(e) => {
-          (e.target as HTMLElement).style.backgroundColor = 'rgb(30, 60, 75)';
-          const img = (e.target as HTMLElement).querySelector('img');
-          if (img) img.style.transform = 'scale(1.1)';
-        }}
-        onMouseUp={(e) => {
-          (e.target as HTMLElement).style.backgroundColor = 'rgb(45, 75, 90)';
-          const img = (e.target as HTMLElement).querySelector('img');
-          if (img) img.style.transform = 'scale(1)';
-        }}
+        onClick={() => onClick()}
+        className='flex items-center justify-center mx-[5px] p-[6px] rounded-lg border-none cursor-pointer bg-[rgb(73,104,118)] hover:bg-[rgb(45,75,90)] text-white flex-1 max-w-[170px] min-w-[170px] text-center transition-all outline-none shadow-none'
       >
-        <img
-          src="/assets/up_circle.svg"
-          alt="Rotated View Icon"
-          style={{
-            width: '18px',
-            height: '18px',
-            marginRight: '7px',
-            transition: 'transform 0.2s ease',
-          }}
+        <Image
+          src='/assets/up_circle.svg'
+          width={18}
+          height={18}
+          className='mr-[7px]'
+          alt='Rotated View Icon'
         />
         Rotated View
       </button>
